refactor(test): extract shared daemon e2e harness + advance_tick helper
The new content_workflow_e2e_test re-declared CollectingLogger /
LiveStatusConsumer / InlineSupervisor / task_folder plus the
status_snapshot status-consumer wiring nearly verbatim from the
pre-existing daemon e2e sibling (~70 lines). Move those five into a shared
`test/support/daemon_e2e_harness.rb` module the new test mixes in, and add
a tiny `advance_tick(dispatcher, supervisor)` helper that names the
`now = Time.now; supervisor.now = now; dispatcher.tick(now: now)` triplet
the drive loop and rest proof both repeated. Behavior-preserving; the
pre-existing sibling is intentionally left untouched (additive per plan
scope).

Hive-Task-Slug: u8-ship-the-content-research-260618-bb97
Hive-Fix-Pass: 03
Hive-Fix-Findings: 3
Hive-Triage-Bias: courageous
Hive-Reviewer-Sources: claude-ce-code-review,codex-ce-code-review,pr-review-toolkit
Hive-Fix-Phase: fix
