+++ test/integration/content_workflow_e2e_test.rb
@@ -1,16 +1,15 @@
 require "test_helper"
-require "json"
 require "shellwords"
 require "hive/cli"
 require "hive/commands/init"
 require "hive/commands/new"
-require "hive/daemon/child_supervisor"
 require "hive/daemon/concurrency_controller"
 require "hive/daemon/dispatcher"
-require "hive/daemon/status_consumer"
+require_relative "../support/daemon_e2e_harness"
 
 class ContentWorkflowE2ETest < Minitest::Test
   include HiveTestHelper
+  include HiveDaemonE2EHarness
 
   TOPIC = "A practical guide to git worktrees for parallel development"
 
@@ -43,9 +42,7 @@ class ContentWorkflowE2ETest < Minitest::Test
           30.times do
             break if File.file?(File.join(task_folder(project_root, "6-done", slug), "article.md"))
 
-            now = Time.now
-            supervisor.now = now
-            dispatcher.tick(now: now)
+            advance_tick(dispatcher, supervisor)
           end
 
           # Rest proof (Decision 2 / R1): once article.md exists the terminal
@@ -55,9 +52,7 @@ class ContentWorkflowE2ETest < Minitest::Test
           # that a regression making it re-dispatchable would be caught.
           ran_at_rest = ran.dup
           spawned_at_rest = supervisor.spawned.size
-          now = Time.now
-          supervisor.now = now
-          dispatcher.tick(now: now)
+          advance_tick(dispatcher, supervisor)
           assert_equal ran_at_rest, ran,
                        "terminal done stage must not re-dispatch an agent after article.md exists"
           assert_equal spawned_at_rest, supervisor.spawned.size,
@@ -100,99 +95,9 @@ class ContentWorkflowE2ETest < Minitest::Test
 
   private
 
-  class CollectingLogger
-    attr_reader :events
-
-    def initialize
-      @events = []
-    end
-
-    def event(name, **attrs)
-      @events << { name: name, attrs: attrs }
-    end
-
-    def close; end
-  end
-
-  class LiveStatusConsumer
-    def initialize(fetch:)
-      @fetch = fetch
-    end
-
-    def fetch
-      @fetch.call
-    end
-  end
-
-  class InlineSupervisor
-    ChildExit = Hive::Daemon::ChildSupervisor::ChildExit
-
-    attr_reader :spawned
-    attr_accessor :now
-
-    def initialize(run:)
-      @run = run
-      @spawned = []
-      @pending = []
-      @next_pid = 7000
-      @now = Time.now
-    end
-
-    def spawn(command_string:, project:, slug:, stage:,
-              hive_state_path: nil, state_file_path: nil, dry_run: nil, request_id: nil)
-      pid = (@next_pid += 1)
-      @spawned << { command: command_string, project: project, slug: slug, stage: stage }
-      exit_code = @run.call(command_string)
-      @pending << ChildExit.new(
-        pid: pid,
-        exit_code: exit_code,
-        project: project,
-        slug: slug,
-        stage: stage,
-        command: command_string,
-        state_file_path: state_file_path,
-        started_at: @now,
-        finished_at: @now,
-        json_envelope: nil
-      )
-      pid
-    end
-
-    def reap_all(now: Time.now)
-      pending = @pending
-      @pending = []
-      pending
-    end
-
-    def reap_dry_run(now: Time.now)
-      []
-    end
-
-    def terminate_all(grace_sec: 600); end
-    def update_timeouts(**); end
-    def enforce_timeouts(now:) = []
-    def in_flight_count = @pending.size
-  end
-
-  def status_snapshot
-    out, = capture_io { Hive::CLI.start([ "status", "--json" ]) }
-    doc = JSON.parse(out)
-    mapper = Hive::Daemon::StatusConsumer.new
-    Hive::Daemon::StatusConsumer::Result.new(
-      ok: true,
-      rows: mapper.send(:extract_rows, doc),
-      projects: mapper.send(:extract_projects, doc),
-      error: nil
-    )
-  end
-
   def content_slug(project_root)
     matches = Dir[File.join(project_root, ".hive-state", "stages", "1-inbox", "*")]
     assert_equal 1, matches.size, "hive new should create exactly one content inbox task"
     File.basename(matches.first)
-  end
-
-  def task_folder(project_root, stage, slug)
-    File.join(project_root, ".hive-state", "stages", stage, slug)
   end
 end
